separate out ligand and protein, add optimiser tests.
