+++ QUBEKit/molecules/ligand.py
@@ -23,7 +23,6 @@ TODO ligand.py Refactor:
         Be more strict about public/private class/method/function naming?
 """
 
-import os
 import pickle
 import xml.etree.ElementTree as ET
 from collections import OrderedDict
@@ -35,10 +34,10 @@ import networkx as nx
 import numpy as np
 import qcelemental as qcel
 
+from QUBEKit.molecules.components import Atom, ExtraSite
+from QUBEKit.molecules.utils import ReadInput
 from QUBEKit.utils import constants
-from QUBEKit.utils.datastructures import Atom, ExtraSite
-from QUBEKit.utils.exceptions import FileTypeError, ConformerError
-from QUBEKit.utils.file_handling import ReadInput, ReadInputProtein
+from QUBEKit.utils.exceptions import ConformerError, FileTypeError
 
 
 class DefaultsMixin:
@@ -243,54 +242,6 @@ class Molecule:
                 return_str += f"\n{key} = {repr(val)}\n"
 
         return return_str
-
-    # def check_names_are_unique(self):
-    #     """
-    #     To prevent problems occurring with some atoms perceived to be the same,
-    #     check the atom names to ensure they are all unique.
-    #     If some are the same, reset all atom names to be: f'{atomic_symbol}{index}'.
-    #     This ensure they are all unique.
-    #     """
-    #
-    #     atom_names = [atom.atom_name for atom in self.atoms]
-    #     # If some atom names aren't unique
-    #     if len(set(atom_names)) < len(atom_names):
-    #         # Change the atom name only; everything else is the same as it was.
-    #         self.atoms = [
-    #             Atom(
-    #                 atomic_number=atom.atomic_number,
-    #                 atom_index=atom.atom_index,
-    #                 atom_name=f'{atom.atomic_symbol}{i}',
-    #                 partial_charge=atom.partial_charge,
-    #                 formal_charge=atom.formal_charge
-    #             )
-    #             for i, atom in enumerate(self.atoms)
-    #         ]
-    #
-    # def _validate_info(self, topology, atoms, coords, input_type, rdkit_molecule=None, descriptors=None):
-    #     """
-    #     Check if the provided information should be stored or not
-    #     :param topology: networkx graph of the topology
-    #     :param atoms: a list of Atom objects
-    #     :param coords: a numpy array of the coords
-    #     :param rdkit_molecule: the rdkit molecule we have extracted the info from
-    #     :param descriptors: a dictionary of the rdkit descriptors
-    #     :return: the updated ligand object
-    #     """
-    #
-    #     # Now check we instancing the ligand if we are then store the info
-    #     if input_type == 'input':
-    #         self.topology = topology
-    #         self.atoms = atoms
-    #         self.descriptors = descriptors
-    #         self.coords[input_type] = coords
-    #         self.rdkit_mol = rdkit_molecule
-    #     else:
-    #         # Check if the new topology is the same then store the new coordinates
-    #         if nx.algorithms.is_isomorphic(self.topology, topology):
-    #             self.coords[input_type] = coords
-    #         else:
-    #             raise TopologyMismatch('Topologies are not the same; cannot store coordinates.')
 
     def get_atom_with_name(self, name):
         """
@@ -950,7 +901,7 @@ class Molecule:
         """Returns a dictionary of atom indices mapped to their class or None if there is no rdkit molecule.
         #TODO we need a to_rdkit method as this should always work for well defined inputs.
         """
-        from QUBEKit.engines import RDKit
+        from QUBEKit.molecules.utils import RDKit
 
         if self.rdkit_mol is not None:
             return RDKit.find_symmetry_classes(self.rdkit_mol)
@@ -1270,137 +1221,3 @@ class Ligand(DefaultsMixin, Molecule):
                     )
 
             pdb_file.write("END\n")
-
-
-class Protein(DefaultsMixin, Molecule):
-    """
-    This class handles the protein input to make the QUBEKit xml files and rewrite the pdb so we can use it.
-    """
-
-    def __init__(self, mol_input, name=None):
-        """
-        is_protein      Bool; True for Protein class
-        home            Current working directory (location for QUBEKit execution).
-        residues        List of all residues in the molecule in order e.g. ['ARG', 'HIS', ... ]
-        Residues        List of residue names for each atom e.g. ['ARG', 'ARG', 'ARG', ... 'HIS', 'HIS', ... ]
-        pdb_names       List
-        """
-
-        super().__init__(mol_input, name)
-
-        self.is_protein = True
-        self.home = os.getcwd()
-        self.residues = None
-        self.Residues = None
-        self.pdb_names = None
-
-        self.combination = "opls"
-
-        if not isinstance(mol_input, ReadInputProtein):
-            self._check_file_type(file_name=mol_input)
-            input_data = ReadInputProtein.from_pdb(file_name=mol_input)
-        else:
-            input_data = mol_input
-        self._save_to_protein(input_data, input_type="input")
-
-    @classmethod
-    def from_file(cls, file_name: str, name: Optional[str] = None) -> "Protein":
-        """
-        Instance the protein class from a pdb file.
-        """
-        cls._check_file_type(file_name=file_name)
-        input_data = ReadInputProtein.from_pdb(file_name=file_name, name=name)
-        return cls(input_data)
-
-    @staticmethod
-    def _check_file_type(file_name: str) -> None:
-        """
-        Make sure the protien is being read from a pdb file.
-        """
-        if ".pdb" not in file_name:
-            raise FileTypeError("Proteins can only be read from pdb.")
-
-    def _save_to_protein(self, mol_input: ReadInputProtein, input_type="input"):
-        """
-        Public access to private file_handlers.py file.
-        Users shouldn't ever need to interface with file_handlers.py directly.
-        All parameters will be set from a file (or other input) via this public method.
-            * Don't bother updating name, topology or atoms if they are already stored.
-            * Do bother updating coords, rdkit_mol, residues, Residues, pdb_names
-        """
-
-        if mol_input.name is not None:
-            self.name = mol_input.name
-        if mol_input.topology is not None:
-            self.topology = mol_input.topology
-        if mol_input.atoms is not None:
-            self.atoms = mol_input.atoms
-        if mol_input.coords is not None:
-            self.coords[input_type] = mol_input.coords
-        if mol_input.residues is not None:
-            self.residues = mol_input.residues
-        if mol_input.pdb_names is not None:
-            self.pdb_names = mol_input.pdb_names
-
-        if not self.topology.edges:
-            print(
-                "No connections found in pdb file; topology will be inferred by OpenMM."
-            )
-            return
-
-        # self.find_angles()
-        # self.find_dihedrals()
-        # self.find_rotatable_dihedrals()
-        # self.find_impropers()
-        # self.measure_dihedrals(input_type)
-        # self.bond_lengths(input_type)
-        # self.measure_angles(input_type)
-        # This creates the dictionary of terms that should be symmetrised.
-        self.symmetrise_from_topology()
-
-    def write_pdb(self, name=None):
-        """
-        This method replaces the ligand method as all of the atom names and residue names have to be replaced.
-        """
-
-        with open(f"{name if name is not None else self.name}.pdb", "w+") as pdb_file:
-
-            pdb_file.write(f"REMARK   1 CREATED WITH QUBEKit {datetime.now()}\n")
-            # Write out the atomic xyz coordinates
-            for i, (coord, atom) in enumerate(zip(self.coords["input"], self.atoms)):
-                x, y, z = coord
-                # May cause issues if protein contains more than 10,000 atoms.
-                pdb_file.write(
-                    f"HETATM {i+1:>4}{atom.atom_name:>5} QUP     1{x:12.3f}{y:8.3f}{z:8.3f}"
-                    f"  1.00  0.00         {atom.atomic_symbol.upper():>3}\n"
-                )
-
-            # Add the connection terms based on the molecule topology.
-            for node in self.topology.nodes:
-                bonded = sorted(list(nx.neighbors(self.topology, node)))
-                if len(bonded) >= 1:
-                    pdb_file.write(
-                        f'CONECT{node + 1:5}{"".join(f"{x + 1:5}" for x in bonded)}\n'
-                    )
-
-            pdb_file.write("END\n")
-
-    def update(self, input_type="input"):
-        """
-        After the protein has been passed to the parametrisation class we get back the bond info
-        use this to update all missing terms.
-        """
-
-        # using the new harmonic bond force dict we can add the bond edges to the topology graph
-        for bond in self.HarmonicBondForce:
-            self.topology.add_edge(*bond)
-
-        # self.find_angles()
-        # self.find_dihedrals()
-        # self.find_rotatable_dihedrals()
-        # self.find_impropers()
-        # self.measure_dihedrals(input_type)
-        # self.bond_lengths(input_type)
-        # self.measure_angles(input_type)
-        # This creates the dictionary of terms that should be symmetrised.
-        self.symmetrise_from_topology()
