+++ QUBEKit/parametrisation/__init__.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env python3
 
 from QUBEKit.utils.helpers import try_load
+
 from .base_parametrisation import Parametrisation
+from .openff import OpenFF
 from .xml import XML
 from .xml_protein import XMLProtein
-from .openff import OpenFF
 
 AnteChamber = try_load("AnteChamber", "QUBEKit.parametrisation.antechamber")
