+++ QUBEKit/proteins/protein_run.py
@@ -6,7 +6,7 @@ import sys
 from functools import partial
 
 from QUBEKit.lennard_jones import LennardJones
-from QUBEKit.ligand import Protein
+from QUBEKit.molecules import Protein
 from QUBEKit.parametrisation import XMLProtein
 from QUBEKit.proteins.protein_tools import get_water, pdb_reformat, qube_general
 from QUBEKit.utils.file_handling import ExtractChargeData
