+++ QUBEKit/tests/protein/parametrisation_test.py
@@ -1,8 +1,8 @@
-from QUBEKit.ligand import Protein
+import shutil
+
+from QUBEKit.molecules import Protein
 from QUBEKit.parametrisation import XMLProtein
 from QUBEKit.utils.file_handling import get_data
-import shutil
-import pytest
 
 
 def test_protein_params(tmpdir):
