+++ QUBEKit/utils/file_handling.py
@@ -11,474 +11,12 @@ TODO
 """
 
 import os
-import re
-from itertools import groupby
-from pathlib import Path
-from typing import List, Optional
 
-import networkx as nx
 import numpy as np
 
-from QUBEKit.engines import RDKit
-from QUBEKit.utils.constants import ANGS_TO_NM, BOHR_TO_ANGS
-from QUBEKit.utils.datastructures import Atom, CustomNamespace, Element, ExtraSite
-
-
-class ReadInput:
-    """
-    Called inside Ligand or Protein; used to handle reading any kind of input valid in QUBEKit
-        QC JSON object
-        SMILES string
-        PDB, MOL2, XYZ file
-    :param mol_input: One of the accepted input types:
-        QC JSON object
-        SMILES string
-        PDB, MOL2, XYZ file
-    :param name: The name of the molecule. Only necessary for smiles strings but can be
-        provided regardless of input type.
-    """
-
-    def __init__(
-        self,
-        topology: Optional[nx.Graph] = None,
-        atoms: Optional[List[Atom]] = None,
-        coords: Optional[np.ndarray] = None,
-        rdkit_mol: Optional = None,
-        name: Optional[str] = None,
-    ):
-
-        # if Path(mol_input).exists():
-        #     self.mol_input = Path(mol_input)
-        #     if name is None:
-        #         self.name = self.mol_input.stem
-        #     else:
-        #         self.name = name
-        # else:
-        #     self.mol_input = mol_input
-        #     self.name = name
-
-        self.topology = topology
-        self.atoms = atoms
-        self.coords = coords
-        self.rdkit_mol = rdkit_mol
-        self.name = name
-
-        # if is_protein:
-        #     self._read_pdb_protein()
-        # else:
-        #     self._read_input()
-
-    @classmethod
-    def from_smiles(cls, smiles: str, name: Optional[str] = None) -> "ReadInput":
-        """
-        Make a ReadInput object which can be taken by the Ligand class to make the model.
-
-        Note
-        ----
-        This method will generate a conformer for the molecule.
-
-        Parameters
-        ----------
-        smiles:
-            The smiles string which should be parsed by rdkit.
-        name:
-            The name that should be given to the molecule.
-        """
-        # Smiles string input
-        rdkit_mol = RDKit.smiles_to_rdkit_mol(smiles_string=smiles, name=name)
-        return cls.from_rdkit(rdkit_mol=rdkit_mol)
-
-    @classmethod
-    def from_file(cls, file_name: str) -> "ReadInput":
-        """
-        Read the input file using RDKit and return the molecule data.
-        """
-        input_file = Path(file_name)
-        # if the file is not there raise an error
-        if not input_file.exists():
-            raise FileNotFoundError(
-                f"{input_file.as_posix()} could not be found is this path correct?"
-            )
-        # xyz is a special case of file only internal readers catch
-        if input_file.suffix == ".xyz":
-            return cls.from_xyz(file_name=input_file.as_posix())
-        # read the input with rdkit
-        rdkit_mol = RDKit.file_to_rdkit_mol(file_path=input_file)
-        return cls.from_rdkit(rdkit_mol=rdkit_mol)
-
-    # def _read_input(self):
-    #     """
-    #     Figure out what the input is (file, smiles, json) and call the relevant method.
-    #     """
-    #
-    #     if self.mol_input.__class__.__name__ == "Molecule":
-    #         # QCArchive object
-    #         self._read_qc_json()
-    #
-    #     elif hasattr(self.mol_input, "stem"):
-    #         # File (pdb, xyz, etc)
-    #         try:
-    #             # Try parse with rdkit:
-    #             self.rdkit_mol = RDKit.mol_input_to_rdkit_mol(self.mol_input)
-    #             self._mol_from_rdkit()
-    #         except AttributeError:
-    #             # Cannot be parsed by rdkit:
-    #             self._read_file()
-    #
-    #     elif isinstance(self.mol_input, str):
-    #         # Smiles string input
-    #         self.rdkit_mol = RDKit.smiles_to_rdkit_mol(self.mol_input, self.name)
-    #         self._mol_from_rdkit()
-    #
-    #     else:
-    #         raise RuntimeError(
-    #             "Cannot read input. mol_input must be a smiles string, path of a file, or qc json."
-    #         )
-
-    # def _read_file(self):
-    #     """
-    #     Called when rdkit cannot parse the file.
-    #     Calls the necessary internal file readers instead.
-    #     """
-    #
-    #     if self.mol_input.suffix == ".pdb":
-    #         self._read_pdb()
-    #     elif self.mol_input.suffix == ".mol2":
-    #         self._read_mol2()
-    #     elif self.mol_input.suffix == ".xyz":
-    #         self._read_xyz()
-    #     else:
-    #         raise FileTypeError(
-    #             f"Could not read file {self.mol_input}. File type must be pdb, mol2 or xyz."
-    #         )
-
-    @classmethod
-    def from_rdkit(cls, rdkit_mol, name: Optional[str] = None) -> "ReadInput":
-        """
-        Using an RDKit Molecule object, extract the name, topology, coordinates and atoms
-        """
-
-        if name is None:
-            name = rdkit_mol.GetProp("_Name")
-
-        atoms = []
-        topology = nx.Graph()
-        # Collect the atom names and bonds
-        for atom in rdkit_mol.GetAtoms():
-            # Collect info about each atom
-            atomic_number = atom.GetAtomicNum()
-            index = atom.GetIdx()
-            try:
-                # PDB file extraction
-                atom_name = atom.GetMonomerInfo().GetName().strip()
-            except AttributeError:
-                try:
-                    # Mol2 file extraction
-                    atom_name = atom.GetProp("_TriposAtomName")
-                except KeyError:
-                    # smiles and mol files have no atom names so generate them here if they are not declared
-                    atom_name = f"{atom.GetSymbol()}{index}"
-
-            qube_atom = Atom(
-                atomic_number, index, atom_name, formal_charge=atom.GetFormalCharge()
-            )
-
-            # Add the atoms as nodes
-            topology.add_node(atom.GetIdx())
-
-            # Add the bonds
-            for bonded in atom.GetNeighbors():
-                topology.add_edge(atom.GetIdx(), bonded.GetIdx())
-                qube_atom.add_bond(bonded.GetIdx())
-
-            # Now add the atom to the molecule
-            atoms.append(qube_atom)
-
-        coords = rdkit_mol.GetConformer().GetPositions()
-        atoms = atoms or None
-        return cls(
-            topology=topology,
-            atoms=atoms,
-            coords=coords,
-            rdkit_mol=rdkit_mol,
-            name=name,
-        )
-
-    @classmethod
-    def from_qc_json(cls, qc_json) -> "ReadInput":
-        """
-        Given a QC JSON object, extracts the topology, atoms and coords of the molecule.
-        """
-
-        topology = nx.Graph()
-        atoms = []
-
-        for i, atom in enumerate(qc_json.symbols):
-            atoms.append(
-                Atom(
-                    atomic_number=Element().number(atom),
-                    atom_index=i,
-                    atom_name=f"{atom}{i}",
-                )
-            )
-            topology.add_node(i)
-
-        for bond in qc_json.connectivity:
-            topology.add_edge(*bond[:2])
-
-        coords = np.array(qc_json.geometry).reshape((len(atoms), 3)) * BOHR_TO_ANGS
-        atoms = atoms or None
-        return cls(topology=topology, atoms=atoms, coords=coords)
-
-    # def _read_pdb(self):
-    #     """
-    #     Internal pdb reader. Only called when RDKit failed to read the pdb.
-    #     Extracts the topology, atoms and coords of the molecule.
-    #     """
-    #
-    #     coords = []
-    #     self.topology = nx.Graph()
-    #     atoms = []
-    #
-    #     atom_count = 0
-    #
-    #     with open(self.mol_input) as pdb:
-    #
-    #         for line in pdb:
-    #             if "ATOM" in line or "HETATM" in line:
-    #                 # start collecting the atom class info
-    #                 atomic_symbol = str(line[76:78])
-    #                 atomic_symbol = re.sub("[0-9]+", "", atomic_symbol)
-    #                 atomic_symbol = atomic_symbol.strip()
-    #                 atom_name = str(line.split()[2])
-    #
-    #                 # If the element column is missing from the pdb, extract the atomic_symbol from the atom name.
-    #                 if not atomic_symbol:
-    #                     atomic_symbol = str(line.split()[2])[:-1]
-    #                     atomic_symbol = re.sub("[0-9]+", "", atomic_symbol)
-    #
-    #                 atomic_number = Element().number(atomic_symbol)
-    #                 # Now instance the qube atom
-    #                 qube_atom = Atom(atomic_number, atom_count, atom_name)
-    #                 atoms.append(qube_atom)
-    #
-    #                 # Also add the atom number as the node in the graph
-    #                 self.topology.add_node(atom_count)
-    #                 atom_count += 1
-    #                 coords.append(
-    #                     [float(line[30:38]), float(line[38:46]), float(line[46:54])]
-    #                 )
-    #
-    #             if "CONECT" in line:
-    #                 atom_index = int(line.split()[1]) - 1
-    #                 # Search the connectivity section and add all edges to the graph corresponding to the bonds.
-    #                 for i in range(2, len(line.split())):
-    #                     if int(line.split()[i]) != 0:
-    #                         bonded_index = int(line.split()[i]) - 1
-    #                         self.topology.add_edge(atom_index, bonded_index)
-    #                         atoms[atom_index].add_bond(bonded_index)
-    #                         atoms[bonded_index].add_bond(atom_index)
-    #
-    #     # put the object back into the correct place
-    #     self.coords = np.array(coords)
-    #     self.atoms = atoms or None
-    #
-    # def _read_mol2(self):
-    #     """
-    #     Internal mol2 reader. Only called when RDKit failed to read the mol2.
-    #     Extracts the topology, atoms and coords of the molecule.
-    #     """
-    #
-    #     coords = []
-    #     self.topology = nx.Graph()
-    #     atoms = []
-    #
-    #     atom_count = 0
-    #
-    #     with open(self.mol_input, "r") as mol2:
-    #
-    #         atom_flag = False
-    #         bond_flag = False
-    #
-    #         for line in mol2:
-    #             if "@<TRIPOS>ATOM" in line:
-    #                 atom_flag = True
-    #                 continue
-    #             elif "@<TRIPOS>BOND" in line:
-    #                 atom_flag = False
-    #                 bond_flag = True
-    #                 continue
-    #             elif "@<TRIPOS>SUBSTRUCTURE" in line:
-    #                 bond_flag = False
-    #                 continue
-    #
-    #             if atom_flag:
-    #                 # Add the molecule information
-    #                 atomic_symbol = line.split()[1][:2]
-    #                 atomic_symbol = re.sub("[0-9]+", "", atomic_symbol)
-    #                 atomic_symbol = atomic_symbol.strip().title()
-    #
-    #                 atomic_number = Element().number(atomic_symbol)
-    #
-    #                 coords.append(
-    #                     [
-    #                         float(line.split()[2]),
-    #                         float(line.split()[3]),
-    #                         float(line.split()[4]),
-    #                     ]
-    #                 )
-    #
-    #                 # Collect the atom names
-    #                 atom_name = str(line.split()[1])
-    #
-    #                 # Add the nodes to the topology object
-    #                 self.topology.add_node(atom_count)
-    #                 atom_count += 1
-    #
-    #                 # Get the atom types
-    #                 atom_type = line.split()[5]
-    #                 atom_type = atom_type.replace(".", "")
-    #
-    #                 # Make the qube_atom
-    #                 qube_atom = Atom(atomic_number, atom_count, atom_name)
-    #                 qube_atom.atom_type = atom_type
-    #
-    #                 atoms.append(qube_atom)
-    #
-    #             if bond_flag:
-    #                 # Add edges to the topology network
-    #                 atom_index, bonded_index = (
-    #                     int(line.split()[1]) - 1,
-    #                     int(line.split()[2]) - 1,
-    #                 )
-    #                 self.topology.add_edge(atom_index, bonded_index)
-    #                 atoms[atom_index].add_bond(bonded_index)
-    #                 atoms[bonded_index].add_bond(atom_index)
-    #
-    #     # put the object back into the correct place
-    #     self.coords = np.array(coords)
-    #     self.atoms = atoms or None
-
-    @classmethod
-    def from_xyz(cls, file_name: str) -> "ReadInput":
-        """
-        Internal xyz reader.
-        Extracts the coords of the molecule.
-        """
-
-        traj_molecules = []
-        coords = []
-
-        with open(file_name) as xyz_file:
-            lines = xyz_file.readlines()
-
-            n_atoms = float(lines[0])
-
-            for line in lines:
-                line = line.split()
-                # skip frame heading lines
-                if len(line) <= 1 or "Iteration" in line:
-                    continue
-
-                coords.append([float(line[1]), float(line[2]), float(line[3])])
-
-                if len(coords) == n_atoms:
-                    # we have collected the molecule now store the frame
-                    traj_molecules.append(np.array(coords))
-                    coords = []
-
-        coords = traj_molecules[0] if len(traj_molecules) == 1 else traj_molecules
-        return cls(coords=coords, topology=None, atoms=None, rdkit_mol=None)
-
-
-class ReadInputProtein:
-    """
-    A class that specialises in reading Protein input files.
-    #TODO are we better doing this with openmm or another tool?
-    """
-
-    def __init__(
-        self,
-        topology: Optional[nx.Graph] = None,
-        atoms: Optional[List[Atom]] = None,
-        coords: Optional[np.ndarray] = None,
-        pdb_names: Optional[List[str]] = None,
-        residues: Optional[List[str]] = None,
-        name: Optional[str] = None,
-    ):
-        self.topology = topology
-        self.atoms = atoms
-        self.coords = coords
-        self.name = name
-        self.residues = residues
-        self.pdb_names = pdb_names
-
-    @classmethod
-    def from_pdb(cls, file_name: str, name: Optional[str] = None):
-        """
-        Read the protein input pdb file.
-        :return:
-        """
-        with open(file_name, "r") as pdb:
-            lines = pdb.readlines()
-
-        coords = []
-        atoms = []
-        topology = nx.Graph()
-        Residues = []
-        pdb_names = []
-
-        # atom counter used for graph node generation
-        atom_count = 0
-        for line in lines:
-            if "ATOM" in line or "HETATM" in line:
-                atomic_symbol = str(line[76:78])
-                atomic_symbol = re.sub("[0-9]+", "", atomic_symbol).strip()
-
-                # If the element column is missing from the pdb, extract the atomic_symbol from the atom name.
-                if not atomic_symbol:
-                    atomic_symbol = str(line.split()[2])
-                    atomic_symbol = re.sub("[0-9]+", "", atomic_symbol)
-
-                # now make sure we have a valid element
-                if atomic_symbol.lower() != "cl" and atomic_symbol.lower() != "br":
-                    atomic_symbol = atomic_symbol[0]
-
-                atom_name = f"{atomic_symbol}{atom_count}"
-                qube_atom = Atom(Element().number(atomic_symbol), atom_count, atom_name)
-
-                atoms.append(qube_atom)
-
-                pdb_names.append(str(line.split()[2]))
-
-                # also get the residue order from the pdb file so we can rewrite the file
-                Residues.append(str(line.split()[3]))
-
-                # Also add the atom number as the node in the graph
-                topology.add_node(atom_count)
-                atom_count += 1
-                coords.append(
-                    [float(line[30:38]), float(line[38:46]), float(line[46:54])]
-                )
-
-            elif "CONECT" in line:
-                conect_terms = line.split()
-                for atom in conect_terms[2:]:
-                    if int(atom):
-                        topology.add_edge(int(conect_terms[1]) - 1, int(atom) - 1)
-
-        coords = np.array(coords)
-        residues = [res for res, group in groupby(Residues)]
-        if name is None:
-            name = Path(file_name).stem
-        return cls(
-            topology=topology,
-            atoms=atoms,
-            coords=coords,
-            pdb_names=pdb_names,
-            residues=residues,
-            name=name,
-        )
+from QUBEKit.molecules.components import ExtraSite
+from QUBEKit.utils.constants import ANGS_TO_NM
+from QUBEKit.utils.datastructures import CustomNamespace
 
 
 class ExtractChargeData:
