+++ QUBEKit/virtual_sites.py
@@ -16,6 +16,7 @@ from matplotlib.cm import ScalarMappable
 from mpl_toolkits.mplot3d import Axes3D
 from scipy.optimize import minimize
 
+from QUBEKit.molecules import ExtraSite
 from QUBEKit.utils.constants import (
     ANGS_TO_NM,
     BOHR_TO_ANGS,
@@ -25,7 +26,6 @@ from QUBEKit.utils.constants import (
     PI,
     VACUUM_PERMITTIVITY,
 )
-from QUBEKit.utils.datastructures import ExtraSite
 from QUBEKit.utils.helpers import append_to_log
